+++ NEWS.md
@@ -1,5 +1,6 @@
 # elixir (development version)
 
+-   elixir now works with formulas.
 -   Removed dependencies on data.table and methods packages.
 
 # elixir 0.1.0